+++ pipda/symbolic.py
@@ -1,6 +1,6 @@
 """Provides Symbolic and Reference classes"""
 from abc import ABC, abstractmethod
-from typing import ClassVar, Any
+from typing import Any
 
 from varname import varname
 
@@ -18,9 +18,6 @@ class Reference(Expression, ABC):
     Args:
         parent: The parent of this reference. For example: `f.A` for `f.A.B`
         ref: The reference. For example: `B` for `f.A.B`
-        context: Defaults to `None`, which should not be specified while
-            instansiation. Because these types of expressions are independent.
-            A context should be passed to `evaluate` to evaluate the expression.
     """
 
     def __init__(self, parent: Any, ref: Any) -> None:
@@ -38,6 +35,16 @@ class Reference(Expression, ABC):
         """Allow `f.a.b` to be a function, so that one can do `f.a.b(1)`"""
         return Function(self, args, kwargs, False)
 
+    @property
+    def _pipda_level(self) -> int:
+        """Get the level of the reference
+
+        Examples:
+            >>> f._pipda_level -> 0
+            >>> f.a._pipda_level -> 1
+        """
+        return getattr(self._pipda_parent, "_pipda_level", 0) + 1
+
     @abstractmethod
     def _pipda_eval(self, data: Any, context: ContextBase = None) -> Any:
         """Evaluate the reference according to the context"""
@@ -51,8 +58,8 @@ class Reference(Expression, ABC):
 class ReferenceAttr(Reference):
     """Attribute references, for example: `f.A`, `f.A.B` etc."""
 
-    # Whether it's a direct reference
-    _pipda_direct: ClassVar[bool] = False
+    def __str__(self) -> str:
+        return f"{self._pipda_parent}.{self._pipda_ref}"
 
     def _pipda_eval(self, data: Any, context: ContextBase = None) -> Any:
         """Evaluate the attribute references"""
@@ -61,15 +68,14 @@ class ReferenceAttr(Reference):
         super()._pipda_eval(data, context)
         parent = evaluate_expr(self._pipda_parent, data, context)
 
-        return context.getattr(
-            parent, self._pipda_ref, self.__class__._pipda_direct
-        )
+        return context.getattr(parent, self._pipda_ref)
 
 
 class ReferenceItem(Reference):
     """Subscript references, for example: `f['A']`, `f.A['B']` etc"""
 
-    _pipda_direct: ClassVar[bool] = False
+    def __str__(self) -> str:
+        return f"{self._pipda_parent}[{self._pipda_ref}]"
 
     def _pipda_eval(self, data: Any, context: ContextBase = None) -> Any:
         """Evaluate the subscript references"""
@@ -77,19 +83,7 @@ class ReferenceItem(Reference):
         parent = evaluate_expr(self._pipda_parent, data, context)
         ref = evaluate_expr(self._pipda_ref, data, context.ref)
 
-        return context.getitem(parent, ref, self.__class__._pipda_direct)
-
-
-class DirectRefAttr(ReferenceAttr):
-    """The direct attribute reference, such as `f.A`"""
-
-    _pipda_direct: ClassVar[bool] = True
-
-
-class DirectRefItem(ReferenceItem):
-    """The direct attribute reference, such as `f['A']`"""
-
-    _pipda_direct: ClassVar[bool] = True
+        return context.getitem(parent, ref)
 
 
 class Symbolic(Expression):
@@ -104,14 +98,22 @@ class Symbolic(Expression):
 
     def __getattr__(self, name: str) -> Any:
         """Create a DirectRefAttr object"""
-        return DirectRefAttr(self, name)
+        return ReferenceAttr(self, name)
 
     def __getitem__(self, item: Any) -> Any:
         """Create a DirectRefItem object"""
-        return DirectRefItem(self, item)
+        return ReferenceItem(self, item)
 
     def __repr__(self) -> str:
         return f"<Symbolic:{self.__name}>"
+
+    def __str__(self) -> str:
+        return self.__name
+
+    @property
+    def _pipda_level(self) -> int:
+        """Level of the reference"""
+        return 0
 
     def _pipda_eval(self, data: Any, context: ContextBase = None) -> Any:
         """When evaluated, this should just return the data directly"""